fix(tests): fix the pollingInterval test
Not sure why this test started failing recently, but it's
failing in main and PR branches in Node 18 and 20 (but not 16
for some reason). fastify is supposed to close idle connections
when `close()` is called, but it's not doing that in the test
for some unknown reason. Setting `forceCloseConnections: true`
makes fastify forcibly close the sockets and shutdown even
if the sockets aren't idle, which seems fine for this test.