Handle gzip compression in response body (#108)
undici request() does not decompress bodies, so the gateway
must do it. This commit only implements gzip decompression
because that is the most common. Note: this commit does
not set `accept-encoding` and that is left the user of the
gateway to decide to enable that by rewriting the request
headers.